ELO and expectation working for ClubAnalysis
