+++ src/player_elo/base_analysis.py
@@ -1,13 +1,18 @@
 # from .elo_calculation_mixin import ELOCalculationMixin
-from .player_elo import GameAnalysis
+from abc import abstractmethod
+from typing import Dict
+
+from .game_analysis import GameAnalysis
+from .database_connection import DatabaseConnection, DATABASE_CONFIG
 
 
 class BaseAnalysis:
     """
     Base class providing common ELO calculation and expectation methods for both Club and Player analysis.
+    @todo: Decide like what exactly this class should do?
     """
 
-    def __init__(self, game_analysis: GameAnalysis, entity_id: int, is_club: bool = True):
+    def __init__(self, game_analysis: GameAnalysis, entity_id: int):
         """
         Initialize.
 
@@ -17,41 +22,30 @@ class BaseAnalysis:
         """
         self.game_analysis = game_analysis
         self.entity_id = entity_id
-        self.is_club = is_club
         self._elo = None
         self._expectation = None
         self._game_score = None
 
-    def fetch_elo(self) -> float:
+    @abstractmethod
+    def _fetch_elo(self) -> float:
         """
         Retrieve or calculate the ELO for the entity (club or player).
         """
-        if self.is_club:
-            # Get rating of CLUB
-            return self.game_analysis.club_ratings.get(self.entity_id, 0)
-        else:
-            # Get rating of PLAYER from player elos with ID, Season
-            self.game_analysis.cur.execute("""
-                SELECT e.elo
-                FROM players_elo e
-                JOIN appearances a ON e.player_id = a.player_id
-                WHERE a.game_id = %s AND e.player_id = %s AND EXTRACT(YEAR FROM a.date::date) = e.season
-            """, (self.game_analysis.game_id, self.entity_id))
-            return self.game_analysis.cur.fetchone()[0]
 
-    def calculate_expectation(self) -> float:
+    @abstractmethod
+    def _calculate_expectation(self) -> float:
         """
         Calculate the expected performance based on ELO ratings.
         """
-        if self.is_club:
-            opponent_id = (
-                self.game_analysis.home_club_id if self.entity_id == self.game_analysis.away_club_id
-                else self.game_analysis.away_club_id)
-            opponent_elo = self.game_analysis.club_ratings[opponent_id]
-        else:
-            opponent_elo = self.game_analysis.club_ratings[self._get_opponent_club_id()]
-
-        return 1 / (1 + pow(10, (opponent_elo - self.elo) / 400))
+        # if self.is_club:
+        #     opponent_id = (
+        #         self.game_analysis.home_club_id if self.entity_id == self.game_analysis.away_club_id
+        #         else self.game_analysis.away_club_id)
+        #     opponent_elo = self.game_analysis.club_ratings[opponent_id]
+        # else:
+        #     opponent_elo = self.game_analysis.club_ratings[self._get_opponent_club_id()]
+        #
+        # return 1 / (1 + pow(10, (opponent_elo - self.elo) / 400))
 
     def update_elo(self, actual_score: float, weight: float) -> float:
         """
@@ -69,45 +63,47 @@ class BaseAnalysis:
         self._elo += change
         return self._elo
 
+    @abstractmethod
     def _get_goal_difference(self) -> int:
         """
         Retrieve goal difference for clubs or match impact for players.
         """
-        if self.is_club:
-            goals_for = len(self.game_analysis.goals_per_club[self.entity_id])
-            opponent_id = (
-                self.game_analysis.home_club_id if self.entity_id == self.game_analysis.away_club_id else self.game_analysis.away_club_id)
-            goals_against = len(self.game_analysis.goals_per_club[opponent_id])
-            return goals_for - goals_against
-        else:
-            return self.game_analysis.match_impact_players[(self._get_club_id(), self.entity_id)]
+        # if self.is_club:
+        #     goals_for = len(self.game_analysis.goals_per_club[self.entity_id])
+        #     opponent_id = (
+        #         self.game_analysis.home_club_id if self.entity_id == self.game_analysis.away_club_id else self.game_analysis.away_club_id)
+        #     goals_against = len(self.game_analysis.goals_per_club[opponent_id])
+        #     return goals_for - goals_against
+        # else:
+        #     return self.game_analysis.match_impact_players[(self._get_club_id(), self.entity_id)]
 
+    @abstractmethod
     def _get_minutes_played(self) -> int:
         """
         Get the minutes played by the player or return full game duration for clubs.
         """
-        if self.is_club:
-            return self.game_analysis.FULL_GAME_MINUTES
-        else:
-            start_min, end_min = self.game_analysis.players_play_times[(self._get_club_id(), self.entity_id)]
-            return end_min - start_min
+        # if self.is_club:
+        #     return self.game_analysis.FULL_GAME_MINUTES
+        # else:
+        #     start_min, end_min = self.game_analysis.players_play_times[(self._get_club_id(), self.entity_id)]
+        #     return end_min - start_min
 
     @property
     def elo(self) -> float:
         if self._elo is None:
-            self._elo = self.fetch_elo()
+            self._elo = self._fetch_elo()
         return self._elo
 
     @property
     def expectation(self) -> float:
         if self._expectation is None:
-            self._expectation = self.calculate_expectation()
+            self._expectation = self._calculate_expectation()
         return self._expectation
 
     @property
     def game_score(self) -> float:
         if self._game_score is None:
-            self._game_score = self.calculate_game_score(analysis=self)
+            self._game_score = self.calculate_game_score()
         return self._game_score
 
     # @staticmethod
@@ -139,6 +135,7 @@ class BaseAnalysis:
                          goal_difference: int = 0, minutes_played: int = 0,
                          minutes_max: int = 90) -> float:
         """
+        @todo NOT DONE YET
         Calculate the change in score.
         """
         res = weight * (game_score - expectation)
@@ -149,27 +146,23 @@ class BaseAnalysis:
         return res
 
 
-class PlayerAnalysis(BaseAnalysis):
-    """
-    Analysis specific to player performance, inheriting shared logic from BaseAnalysis.
-    """
-
-    def __init__(self, game_analysis: GameAnalysis, player_id: int):
-        super().__init__(game_analysis, entity_id=player_id, is_club=False)
-
-    def _get_club_id(self) -> int:
+    def summary(self) -> Dict[str, any]:
         """
-        Retrieve the club ID for the player.
+        Generate a summary of key attributes and properties for easy viewing during testing.
+
+        @return: Dictionary summarizing the GameAnalysis instance.
         """
-        self.game_analysis.cur.execute("""
-                SELECT player_club_id
-                FROM appearances
-                WHERE game_id = %s AND player_id = %s
-            """, (self.game_analysis.game_id, self.entity_id))
-        return self.game_analysis.cur.fetchone()[0]
+        return {
+            "elo": self.elo,
+            'expectation': self.expectation,
+            'game_score': self.game_score,
+        }
 
+    def print_summary(self):
+        """
+        Print a formatted summary of key attributes and properties for easy testing and debugging.
+        """
+        summary = self.summary()
+        for key, value in summary.items():
+            print(f"{key}: {value}")
 
-class ClubAnalysis(BaseAnalysis):
-
-    def __init__(self, game_analysis: GameAnalysis, club_id: int):
-        super().__init__(game_analysis, entity_id=club_id, is_club=True)
