+++ src/player_elo/database_connection.py
@@ -1,0 +1,29 @@
+from typing import Dict
+
+import psycopg
+
+DATABASE_CONFIG = {
+    'dbname': 'football',
+    'user': 'postgres',
+    'password': '1234',
+    'host': 'localhost',
+    'port': '5432'
+}
+
+
+class DatabaseConnection:
+    """
+    Set up connection with PostgreSQL database
+    """
+
+    def __init__(self, config: Dict[str, str]):
+        self.config = config
+        self.conn = None
+
+    def __enter__(self):
+        self.conn = psycopg.connect(**self.config)
+        return self.conn
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        if self.conn:
+            self.conn.close()
