+++ src/player_elo/player_analysis.py
@@ -1,5 +1,6 @@
 from src.player_elo.base_analysis import BaseAnalysis
-from src.player_elo.player_elo import GameAnalysis
+from src.player_elo.database_connection import DatabaseConnection, DATABASE_CONFIG
+from src.player_elo.game_analysis import GameAnalysis
 
 
 class PlayerAnalysis(BaseAnalysis):
@@ -8,7 +9,29 @@ class PlayerAnalysis(BaseAnalysis):
     """
 
     def __init__(self, game_analysis: GameAnalysis, player_id: int):
-        super().__init__(game_analysis, entity_id=player_id, is_club=False)
+        super().__init__(game_analysis, entity_id=player_id)
+        # IDK maybe delete this cuz it's confusing?
+        self.player_id = player_id
+
+    def _fetch_elo(self) -> float:
+        self.game_analysis.cur.execute("""
+                        SELECT e.elo
+                        FROM players_elo e
+                        JOIN appearances a ON e.player_id = a.player_id
+                        WHERE a.game_id = %s AND e.player_id = %s AND EXTRACT(YEAR FROM a.date::date) = e.season
+                    """, (self.game_analysis.game_id, self.entity_id))
+        return self.game_analysis.cur.fetchone()[0]
+
+    def _calculate_expectation(self) -> float:
+        opponent_elo = self.game_analysis.club_ratings[self._get_opponent_club_id()]
+        return 1 / (1 + pow(10, (opponent_elo - self.elo) / 400))
+
+    def _get_goal_difference(self) -> int:
+        return self.game_analysis.match_impact_players[(self._get_club_id(), self.entity_id)]
+
+    def _get_minutes_played(self) -> int:
+        start_min, end_min = self.game_analysis.players_play_times[(self._get_club_id(), self.entity_id)]
+        return end_min - start_min
 
     def _get_club_id(self) -> int:
         """
@@ -25,4 +48,53 @@ class PlayerAnalysis(BaseAnalysis):
 class ClubAnalysis(BaseAnalysis):
 
     def __init__(self, game_analysis: GameAnalysis, club_id: int):
-        super().__init__(game_analysis, entity_id=club_id, is_club=True)
+        super().__init__(game_analysis, entity_id=club_id)
+
+    def _fetch_elo(self) -> float:
+        return self.game_analysis.club_ratings.get(self.entity_id, 0)
+
+    def _calculate_expectation(self) -> float:
+        opponent_id = (
+            self.game_analysis.home_club_id if self.entity_id == self.game_analysis.away_club_id
+            else self.game_analysis.away_club_id)
+        opponent_elo = self.game_analysis.club_ratings[opponent_id]
+        return 1 / (1 + pow(10, (opponent_elo - self.elo) / 400))
+
+    def _get_goal_difference(self) -> int:
+        goals_for = len(self.game_analysis.goals_per_club[self.entity_id])
+        opponent_id = (
+            self.game_analysis.home_club_id if self.entity_id == self.game_analysis.away_club_id else self.game_analysis.away_club_id)
+        goals_against = len(self.game_analysis.goals_per_club[opponent_id])
+        return goals_for - goals_against
+
+    def _get_minutes_played(self) -> int:
+        return self.game_analysis.FULL_GAME_MINUTES
+
+
+with DatabaseConnection(DATABASE_CONFIG) as conn:
+    with conn.cursor() as cur:
+        # Initialize game analysis with game ID
+        game_analysis = GameAnalysis(cur, game_id=3079452)
+
+        # Initialize BaseAnalysis for a club entity
+        home_club_analysis = ClubAnalysis(game_analysis, game_analysis.home_club_id)
+        away_club_analysis = ClubAnalysis(game_analysis, game_analysis.away_club_id)
+        # Calculate ELO and expectation
+        print("Club ELO:", home_club_analysis.elo)
+        print("Club ELO:", away_club_analysis.elo)
+        print("Club Expectation:", home_club_analysis.expectation)
+        print("Club Expectation:", away_club_analysis.expectation)
+        # Update the club's ELO based on actual game score (e.g., actual_score=1.0 if they won)
+        # updated_elo = club_analysis.update_elo(actual_score=1.0, weight=0.5)
+        # print("Updated Club ELO:", updated_elo)
+
+        # Initialize BaseAnalysis for a player entity
+        # player_analysis = BaseAnalysis(game_analysis=game_analysis, entity_id=20506, is_club=False)
+        #
+        # # Calculate ELO and expectation for the player
+        # print("Player ELO:", player_analysis.elo)
+        # print("Player Expectation:", player_analysis.expectation)
+        #
+        # # Update the player's ELO based on actual performance score
+        # updated_player_elo = player_analysis.update_elo(actual_score=0.5, weight=0.3)
+        # print("Updated Player ELO:", updated_player_elo)
